requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
supabase
bcrypt
pyjwt
email-validator
python-multipart
aiofiles